Extract Pager component from Hotspots
REDMINE-20853
