+++ entry_types/scrolled/package/src/contentElements/hotspots/PagerButton.module.css
